backend/requirements.txt
# ─── Core framework ───────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# ─── Database ─────────────────────────────────────────────────
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
psycopg2-binary==2.9.9

# ─── LlamaIndex ───────────────────────────────────────────────
llama-index-core==0.10.68
llama-index-vector-stores-postgres==0.1.14
llama-index-embeddings-openai==0.1.11
llama-index-llms-anthropic==0.1.13
llama-index-readers-file==0.1.33

# ─── Embeddings & reranker ────────────────────────────────────
openai==1.40.0
sentence-transformers==3.0.1    # free cross-encoder reranker

# ─── LLM ─────────────────────────────────────────────────────
anthropic==0.28.1

# ─── Document processing ──────────────────────────────────────
pypdf==4.3.1
langdetect==1.0.9

# ─── Auth & security ──────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Dev & testing ────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
